Fix: Use shared 'ansible_extra_vars' function
Fixes #3204
